backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
PyJWT==2.12.1
psycopg2-binary==2.9.9
